+++ NEWS.md
@@ -8,3 +8,9 @@
 * Updated code and tests to run with updated versions of the mice and ggplot2 packages
 * Updated documentation of package man page to include features not supported at the moment
 
+# adjustedCurves 0.10.0
+
+* Removed support for tmle, ostmle methods
+* Changed citation information because manuscript was published
+* Changed `print` method to be equal to `summary` method
+* Fixed issues with unit-tests that require packages under "Suggests" only